+++ temba/api/v1/serializers.py
@@ -205,7 +205,7 @@ class ContactReadSerializer(ReadSerializer):
 
         for contact_field in self.context['contact_fields']:
             value = obj.get_field(contact_field.key)
-            fields[contact_field.key] = Contact.serialize_field_value(contact_field, value)
+            fields[contact_field.key] = Contact.serialize_field_value_legacy(contact_field, value)
         return fields
 
     def get_tel(self, obj):
@@ -526,8 +526,6 @@ class FlowRunReadSerializer(ReadSerializer):
     completed = serializers.SerializerMethodField('is_completed')
     created_on = DateTimeField()
     modified_on = DateTimeField()
-    expires_on = DateTimeField()
-    expired_on = serializers.SerializerMethodField()
 
     def get_flow_uuid(self, obj):
         return obj.flow.uuid
@@ -538,12 +536,9 @@ class FlowRunReadSerializer(ReadSerializer):
     def is_completed(self, obj):
         return obj.is_completed()
 
-    def get_expired_on(self, obj):
-        return format_datetime(obj.exited_on) if obj.exit_type == FlowRun.EXIT_TYPE_EXPIRED else None
-
     class Meta:
         model = FlowRun
-        fields = ('flow_uuid', 'run', 'contact', 'completed', 'created_on', 'modified_on', 'expires_on', 'expired_on')
+        fields = ('flow_uuid', 'run', 'contact', 'completed', 'created_on', 'modified_on',)
 
 
 class FlowRunWriteSerializer(WriteSerializer):